Implmented ViewPager with Turbolinks and BottomNav
Turbolinks sessions are stored in the TurbolinksSessionManager and loaded into the 4 views managed by Viewpager. ScoutPagerAdapter handles that integration. The ViewPager is then loaded alongside a BottomNav. The two are synchronized by work done within the activity itself, it is set up to receive callbacks from each class and update the other accordingly